requirements.txt
streamlit
openai
anthropic
google-generativeai
python-pptx
